Minor changes to allow compilation on Ubuntu Breezy amd64. Make the configure script test for, and require lame to be installed. Fix the test for libdvdnav to also check for the header.
git-svn-id: http://svn.mythtv.org/svn/trunk@6741 7dbf422c-18fa-0310-86e9-fd20926502f2

+++ mythtv/libs/libmythtv/signalmonitor.cpp
@@ -82,7 +82,7 @@ SignalMonitor *SignalMonitor::Init(QString cardtype, int db_cardnum,
     {
         VERBOSE(VB_IMPORTANT,
                 QString("Failed to create signal monitor in Init(%1, %2, 0x%3)")
-                .arg(cardtype).arg(db_cardnum).arg((int)channel,0,16));
+                .arg(cardtype).arg(db_cardnum).arg((long)channel,0,16));
     }
 
     return signalMonitor;
